Updated TabbedContent Pt2 and tests
