+++ src/stories/components/TabbedContent.stories.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
 import Wrapper from '../../components/Wrapper';
 import Card from '../../components/Card';
-import ActionTabbedContent, {
-  ActionTabbedContentProps,
-} from '../../components/ActionTabbedContent';
+import TabbedContent, {
+  TabbedContentProps,
+} from '../../components/TabbedContent';
 import { COLORS } from '../../common/styles/CMTheme';
 import { TAction } from '../../types/Actions';
 
@@ -20,7 +20,7 @@ const dummyAction: TAction = {
 
 export default {
   title: 'ClimateMind/components/TabbedContent',
-  component: ActionTabbedContent,
+  component: TabbedContent,
   decorators: [
     (Story) => (
       <Wrapper bgColor={COLORS.SECONDARY} fullHeight>
@@ -32,8 +32,8 @@ export default {
   ],
 } as Meta;
 
-const Template: Story<ActionTabbedContentProps> = (args) => (
-  <ActionTabbedContent {...args} />
+const Template: Story<TabbedContentProps> = (args) => (
+  <TabbedContent {...args} />
 );
 
 export const Default = Template.bind({});
